third_party/blink/web_tests/external/wpt/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/legend-display-expected.txt

This is a testharness.js-based test.
Found 76 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] rendered legend with display: table-row-group
  assert_equals: display: table-row-group is not supported expected "table-row-group" but got "block"
[FAIL] rendered legend with display: table-header-group
  assert_equals: display: table-header-group is not supported expected "table-header-group" but got "block"
[FAIL] rendered legend with display: table-footer-group
  assert_equals: display: table-footer-group is not supported expected "table-footer-group" but got "block"
[FAIL] rendered legend with display: table-row
  assert_equals: display: table-row is not supported expected "table-row" but got "block"
[FAIL] rendered legend with display: table-cell
  assert_equals: display: table-cell is not supported expected "table-cell" but got "block"
[FAIL] rendered legend with display: table-column-group
  assert_equals: display: table-column-group is not supported expected "table-column-group" but got "block"
[FAIL] rendered legend with display: table-column
  assert_equals: display: table-column is not supported expected "table-column" but got "block"
[FAIL] rendered legend with display: table-caption
  assert_equals: display: table-caption is not supported expected "table-caption" but got "block"
[FAIL] rendered legend with display: run-in
  assert_equals: display: run-in is not supported expected "run-in" but got "block"
[FAIL] rendered legend with display: inline
  assert_equals: display: inline is not supported expected "inline" but got "block"
[FAIL] rendered legend with display: inline-block
  assert_equals: display: inline-block is not supported expected "inline-block" but got "block"
[FAIL] rendered legend with display: inline-table
  assert_equals: display: inline-table is not supported expected "inline-table" but got "table"
[FAIL] rendered legend with display: ruby
  assert_equals: display: ruby is not supported expected "ruby" but got "block ruby"
[FAIL] rendered legend with display: ruby-base
  assert_equals: display: ruby-base is not supported expected "ruby-base" but got "block"
[FAIL] rendered legend with display: ruby-text
  assert_equals: display: ruby-text is not supported expected "ruby-text" but got "block"
[FAIL] rendered legend with display: ruby-base-container
  assert_equals: display: ruby-base-container is not supported expected "ruby-base-container" but got "block"
[FAIL] rendered legend with display: ruby-text-container
  assert_equals: display: ruby-text-container is not supported expected "ruby-text-container" but got "block"
[FAIL] rendered legend with display: inline-grid
  assert_equals: display: inline-grid is not supported expected "inline-grid" but got "grid"
[FAIL] rendered legend with display: inline-flex
  assert_equals: display: inline-flex is not supported expected "inline-flex" but got "flex"
[FAIL] rendered legend with display: table-row-group; overflow:hidden
  assert_equals: display: table-row-group is not supported expected "table-row-group" but got "block"
[FAIL] rendered legend with display: table-header-group; overflow:hidden
  assert_equals: display: table-header-group is not supported expected "table-header-group" but got "block"
[FAIL] rendered legend with display: table-footer-group; overflow:hidden
  assert_equals: display: table-footer-group is not supported expected "table-footer-group" but got "block"
[FAIL] rendered legend with display: table-row; overflow:hidden
  assert_equals: display: table-row is not supported expected "table-row" but got "block"
[FAIL] rendered legend with display: table-cell; overflow:hidden
  assert_equals: display: table-cell is not supported expected "table-cell" but got "block"
[FAIL] rendered legend with display: table-column-group; overflow:hidden
  assert_equals: display: table-column-group is not supported expected "table-column-group" but got "block"
[FAIL] rendered legend with display: table-column; overflow:hidden
  assert_equals: display: table-column is not supported expected "table-column" but got "block"
[FAIL] rendered legend with display: table-caption; overflow:hidden
  assert_equals: display: table-caption is not supported expected "table-caption" but got "block"
[FAIL] rendered legend with display: run-in; overflow:hidden
  assert_equals: display: run-in is not supported expected "run-in" but got "block"
[FAIL] rendered legend with display: inline; overflow:hidden
  assert_equals: display: inline is not supported expected "inline" but got "block"
[FAIL] rendered legend with display: inline-block; overflow:hidden
  assert_equals: display: inline-block is not supported expected "inline-block" but got "block"
[FAIL] rendered legend with display: inline-table; overflow:hidden
  assert_equals: display: inline-table is not supported expected "inline-table" but got "table"
[FAIL] rendered legend with display: ruby; overflow:hidden
  assert_equals: display: ruby is not supported expected "ruby" but got "block ruby"
[FAIL] rendered legend with display: ruby-base; overflow:hidden
  assert_equals: display: ruby-base is not supported expected "ruby-base" but got "block"
[FAIL] rendered legend with display: ruby-text; overflow:hidden
  assert_equals: display: ruby-text is not supported expected "ruby-text" but got "block"
[FAIL] rendered legend with display: ruby-base-container; overflow:hidden
  assert_equals: display: ruby-base-container is not supported expected "ruby-base-container" but got "block"
[FAIL] rendered legend with display: ruby-text-container; overflow:hidden
  assert_equals: display: ruby-text-container is not supported expected "ruby-text-container" but got "block"
[FAIL] rendered legend with display: inline-grid; overflow:hidden
  assert_equals: display: inline-grid is not supported expected "inline-grid" but got "grid"
[FAIL] rendered legend with display: inline-flex; overflow:hidden
  assert_equals: display: inline-flex is not supported expected "inline-flex" but got "flex"
[FAIL] rendered legend with display: table-row-group; columns:1
  assert_equals: display: table-row-group is not supported expected "table-row-group" but got "block"
[FAIL] rendered legend with display: table-header-group; columns:1
  assert_equals: display: table-header-group is not supported expected "table-header-group" but got "block"
[FAIL] rendered legend with display: table-footer-group; columns:1
  assert_equals: display: table-footer-group is not supported expected "table-footer-group" but got "block"
[FAIL] rendered legend with display: table-row; columns:1
  assert_equals: display: table-row is not supported expected "table-row" but got "block"
[FAIL] rendered legend with display: table-cell; columns:1
  assert_equals: display: table-cell is not supported expected "table-cell" but got "block"
[FAIL] rendered legend with display: table-column-group; columns:1
  assert_equals: display: table-column-group is not supported expected "table-column-group" but got "block"
[FAIL] rendered legend with display: table-column; columns:1
  assert_equals: display: table-column is not supported expected "table-column" but got "block"
[FAIL] rendered legend with display: table-caption; columns:1
  assert_equals: display: table-caption is not supported expected "table-caption" but got "block"
[FAIL] rendered legend with display: run-in; columns:1
  assert_equals: display: run-in is not supported expected "run-in" but got "block"
[FAIL] rendered legend with display: inline; columns:1
  assert_equals: display: inline is not supported expected "inline" but got "block"
[FAIL] rendered legend with display: inline-block; columns:1
  assert_equals: display: inline-block is not supported expected "inline-block" but got "block"
[FAIL] rendered legend with display: inline-table; columns:1
  assert_equals: display: inline-table is not supported expected "inline-table" but got "table"
[FAIL] rendered legend with display: ruby; columns:1
  assert_equals: display: ruby is not supported expected "ruby" but got "block ruby"
[FAIL] rendered legend with display: ruby-base; columns:1
  assert_equals: display: ruby-base is not supported expected "ruby-base" but got "block"
[FAIL] rendered legend with display: ruby-text; columns:1
  assert_equals: display: ruby-text is not supported expected "ruby-text" but got "block"
[FAIL] rendered legend with display: ruby-base-container; columns:1
  assert_equals: display: ruby-base-container is not supported expected "ruby-base-container" but got "block"
[FAIL] rendered legend with display: ruby-text-container; columns:1
  assert_equals: display: ruby-text-container is not supported expected "ruby-text-container" but got "block"
[FAIL] rendered legend with display: inline-grid; columns:1
  assert_equals: display: inline-grid is not supported expected "inline-grid" but got "grid"
[FAIL] rendered legend with display: inline-flex; columns:1
  assert_equals: display: inline-flex is not supported expected "inline-flex" but got "flex"
[FAIL] rendered legend with display: table-row-group; overflow:hidden;columns:1
  assert_equals: display: table-row-group is not supported expected "table-row-group" but got "block"
[FAIL] rendered legend with display: table-header-group; overflow:hidden;columns:1
  assert_equals: display: table-header-group is not supported expected "table-header-group" but got "block"
[FAIL] rendered legend with display: table-footer-group; overflow:hidden;columns:1
  assert_equals: display: table-footer-group is not supported expected "table-footer-group" but got "block"
[FAIL] rendered legend with display: table-row; overflow:hidden;columns:1
  assert_equals: display: table-row is not supported expected "table-row" but got "block"
[FAIL] rendered legend with display: table-cell; overflow:hidden;columns:1
  assert_equals: display: table-cell is not supported expected "table-cell" but got "block"
[FAIL] rendered legend with display: table-column-group; overflow:hidden;columns:1
  assert_equals: display: table-column-group is not supported expected "table-column-group" but got "block"
[FAIL] rendered legend with display: table-column; overflow:hidden;columns:1
  assert_equals: display: table-column is not supported expected "table-column" but got "block"
[FAIL] rendered legend with display: table-caption; overflow:hidden;columns:1
  assert_equals: display: table-caption is not supported expected "table-caption" but got "block"
[FAIL] rendered legend with display: run-in; overflow:hidden;columns:1
  assert_equals: display: run-in is not supported expected "run-in" but got "block"
[FAIL] rendered legend with display: inline; overflow:hidden;columns:1
  assert_equals: display: inline is not supported expected "inline" but got "block"
[FAIL] rendered legend with display: inline-block; overflow:hidden;columns:1
  assert_equals: display: inline-block is not supported expected "inline-block" but got "block"
[FAIL] rendered legend with display: inline-table; overflow:hidden;columns:1
  assert_equals: display: inline-table is not supported expected "inline-table" but got "table"
[FAIL] rendered legend with display: ruby; overflow:hidden;columns:1
  assert_equals: display: ruby is not supported expected "ruby" but got "block ruby"
[FAIL] rendered legend with display: ruby-base; overflow:hidden;columns:1
  assert_equals: display: ruby-base is not supported expected "ruby-base" but got "block"
[FAIL] rendered legend with display: ruby-text; overflow:hidden;columns:1
  assert_equals: display: ruby-text is not supported expected "ruby-text" but got "block"
[FAIL] rendered legend with display: ruby-base-container; overflow:hidden;columns:1
  assert_equals: display: ruby-base-container is not supported expected "ruby-base-container" but got "block"
[FAIL] rendered legend with display: ruby-text-container; overflow:hidden;columns:1
  assert_equals: display: ruby-text-container is not supported expected "ruby-text-container" but got "block"
[FAIL] rendered legend with display: inline-grid; overflow:hidden;columns:1
  assert_equals: display: inline-grid is not supported expected "inline-grid" but got "grid"
[FAIL] rendered legend with display: inline-flex; overflow:hidden;columns:1
  assert_equals: display: inline-flex is not supported expected "inline-flex" but got "flex"
Harness: the test ran to completion.
